+++ CHANGES.md
@@ -4,6 +4,10 @@ Changelog
 In Development
 --------------
 
+- `LatinBaseValidator` ora indica quali sono i caratteri non accettati ([#185][185]).
+
+[185]: https://github.com/FatturaElettronica/FatturaElettronica.NET/issues/185
+
 v2.0.3
 ------